# backend/openapi.yaml
components:
  schemas:
    controller.DecommissionRequest:
      properties:
        decommissioned_date:
          type: string
        notes:
          type: string
      type: object
    controller.ReactivateRequest:
      properties:
        notes:
          type: string
      type: object
    controller.apiResponse-any:
      properties:
        data: {}
        error:
          type: string
      type: object
...
    controller.apiResponse-array_model_FismaSystem:
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.FismaSystem'
          type: array
          uniqueItems: false
        error:
          type: string
      type: object
...
    controller.apiResponse-array_model_FunctionOption:
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.FunctionOption'
          type: array
          uniqueItems: false
        error:
          type: string
      type: object
...
    controller.apiResponse-model_FismaSystem:
      properties:
        data:
          $ref: '#/components/schemas/model.FismaSystem'
        error:
          type: string
      type: object
...
    model.FismaSystem:
      properties:
        component:
          type: string
        datacallcontact:
          type: string
        datacenterenvironment:
          type: string
        decommissioned:
          type: boolean
        decommissioned_by:
          type: string
        decommissioned_date:
          type: string
        decommissioned_notes:
          type: string
        divisionname:
          type: string
        fismaacronym:
          type: string
        fismaname:
          type: string
        fismasubsystem:
          type: string
        fismasystemid:
          type: integer
        fismauid:
          type: string
        groupacronym:
          type: string
        groupname:
          type: string
        issoemail:
          type: string
        opdiv_id:
          type: integer
        reactivated_by:
          type: string
        reactivated_date:
          type: string
        reactivation_notes:
          type: string
        sdl_sync_enabled:
          type: boolean
      type: object
...
    model.FunctionOption:
      properties:
        description:
          type: string
        functionid:
          type: integer
        functionoptionid:
          type: integer
        optionname:
          type: string
        score:
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: API for the Zero Trust Maturity Framework (ZTMF).
  title: ZTMF API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /fismasystems:
    get:
      parameters:
      - description: Filter by decommissioned status
        in: query
        name: decommissioned
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_model_FismaSystem'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List all FISMA systems
      tags:
      - fismasystems
    post:
      parameters:
      - description: FISMA system ID (update only)
        in: path
        name: fismasystemid
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.FismaSystem'
                description: FISMA system to save
                summary: body
        description: FISMA system to save
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_FismaSystem'
          description: Created
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a FISMA system
      tags:
      - fismasystems
  /fismasystems/{fismasystemid}:
    delete:
      parameters:
      - description: FISMA system ID
        in: path
        name: fismasystemid
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/controller.DecommissionRequest'
                description: Optional decommission parameters
                summary: body
        description: Optional decommission parameters
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Decommission a FISMA system
      tags:
      - fismasystems
    get:
      parameters:
      - description: FISMA system ID
        in: path
        name: fismasystemid
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_FismaSystem'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get a FISMA system by ID
      tags:
      - fismasystems
    put:
      parameters:
      - description: FISMA system ID (update only)
        in: path
        name: fismasystemid
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.FismaSystem'
                description: FISMA system to save
                summary: body
        description: FISMA system to save
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_FismaSystem'
          description: Created
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a FISMA system
      tags:
      - fismasystems
...
  /fismasystems/{fismasystemid}/reactivate:
    put:
      parameters:
      - description: FISMA system ID
        in: path
        name: fismasystemid
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/controller.ReactivateRequest'
                description: Optional reactivation parameters
                summary: body
        description: Optional reactivation parameters
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_FismaSystem'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Reactivate a decommissioned FISMA system
      tags:
      - fismasystems
...
  /functions/{functionid}/options:
    get:
      parameters:
      - description: Function ID
        in: path
        name: functionid
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_model_FunctionOption'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List options for a function
      tags:
      - functions
...
servers:
- url: /api/v1